inference_submission/requirements.txt
# Core ML/AI packages
torch
torchvision
timm
numpy<2.0  # Pin to avoid compatibility issues with PyTorch

# Image processing
SimpleITK
pyvips
Pillow
tifffile
imagecodecs
opencv-python  # Required for cv2 in WholeSlideImage.py
openslide-python  # Required for openslide in extract_features_fp.py and WholeSlideImage.py

# Data processing
pandas  # Required for DataFrame operations
h5py  # Required for HDF5 file operations in dataset_h5.py and WholeSlideImage.py

# Stain normalization (for feature extraction)
torchstain  # Required for stain normalization in dataset_h5.py

# System utilities
psutil
